+++ src/main/java/umc/meme/auth/global/auth/AuthService.java
@@ -14,8 +14,8 @@ import umc.meme.auth.domain.token.entity.Token;
 import umc.meme.auth.domain.token.entity.TokenRepository;
 import umc.meme.auth.domain.user.entity.User;
 import umc.meme.auth.domain.user.entity.UserRepository;
-import umc.meme.auth.global.auth.converter.TokenConverter;
-import umc.meme.auth.global.auth.converter.UserConverter;
+import umc.meme.auth.global.converter.TokenConverter;
+import umc.meme.auth.global.converter.UserConverter;
 import umc.meme.auth.global.auth.dto.AuthRequest;
 import umc.meme.auth.global.auth.dto.AuthResponse;
 import umc.meme.auth.global.enums.Provider;
